+++ tests/Unit/Internal/Domain/Contact/From/ContactFormEmailValidatorTest.php
@@ -9,7 +9,7 @@ declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Tests\Unit\Internal\Domain\Contact\Form;
 
-use OxidEsales\EshopCommunity\Internal\Domain\Email\EmailValidatorService;
+use OxidEsales\EshopCommunity\Internal\Utility\Email\EmailValidatorService;
 use OxidEsales\EshopCommunity\Internal\Framework\Form\Form;
 use OxidEsales\EshopCommunity\Internal\Framework\Form\FormField;
 use OxidEsales\EshopCommunity\Internal\Domain\Contact\Form\ContactFormEmailValidator;